+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This document outlines major changes between releases.
 ### Removed
 
 ### Updated
+- NeoFS SDK to v1.0.0-rc.13.0.20250715070617-c7038b450691 (API 2.18, #311)
 
 ## Upgrading from 0.12.0
 - Rename `pool.container_ops_timeout` to `pool.container-ops-timeout` if used